requirements.txt
# Discord Bot
discord.py>=2.3.0

# LLM APIs
openai>=1.3.0

# Search API & HTTP Client
aiohttp>=3.8.0
httpx[http2]>=0.24.0

# NLP & ML
transformers>=4.30.0
sentence-transformers>=2.2.0
torch>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
alembic>=1.11.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
pydantic>=2.0.0
click>=8.1.0
rich>=13.0.0
tenacity>=8.2.0

# Date & Time
python-dateutil>=2.8.0
pytz>=2023.3

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# File Processing
PyPDF2>=3.0.0
python-docx>=0.8.11
openpyxl>=3.1.0

# Security & Validation
cryptography>=41.0.0
validators>=0.20.0

# Performance & Monitoring
psutil>=5.9.0
memory-profiler>=0.61.0

# API Rate Limiting
ratelimit>=2.2.1
asyncio-throttle>=1.0.2

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
pytest-xdist>=3.3.0
factory-boy>=3.3.0